+++ include/managers/settings_manager.h
@@ -91,6 +91,7 @@ typedef struct {
   
   int32_t esp_comm_tx_pin; // ESP communication TX pin
   int32_t esp_comm_rx_pin; // ESP communication RX pin
+  bool ap_enabled; // Enable/disable AP across reboots
 } FSettings;
 
 // Function declarations
@@ -198,6 +199,10 @@ bool settings_get_web_auth_enabled(const FSettings *settings);
 
 void settings_set_esp_comm_pins(FSettings *settings, int32_t tx_pin, int32_t rx_pin);
 void settings_get_esp_comm_pins(const FSettings *settings, int32_t *tx_pin, int32_t *rx_pin);
+
+// Getter and Setter for AP enabled state
+void settings_set_ap_enabled(FSettings *settings, bool enabled);
+bool settings_get_ap_enabled(const FSettings *settings);
 
 extern FSettings G_Settings;